Un-eject react app - to upgrade dependencies
We ejected, ages ago, mainly to allow pdf generation.

However upgrading all the dependencies is a total pain, particularly the
substantial webpack changes.

So we un-eject, which is a tiny bit messy, but not as bad as not.

Also figured out how to support pdf generation within the
create-react-app ecosystem, not sure if this was possible before.